database quikchanges

+++ migrations/20240620064644-create_business_Card.js
@@ -47,7 +47,7 @@ module.exports = {
         type: Sequelize.BOOLEAN
       },
       mobileNumber: {
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(20)
       },
       companyName: {
         type: Sequelize.STRING
@@ -56,7 +56,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       whatsapp: {
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(20)
       },
       facebook: {
         type: Sequelize.STRING
